added auto race complete service and proper functionality for completed races

+++ RunningClub/Models/AppUser.cs
@@ -7,7 +7,7 @@ public class AppUser : IdentityUser
 {
     public required string FName { get; set; }
     public required string LName { get; set; }
-    public float Mileage { get; set; }
+    public double Mileage { get; set; }
     public List<MemberClub> Clubs { get; set; } = new();
     public List<MemberRace> Races { get; set; } = new();
